test(state-sync): fix shadow_tracking under P2P sync; move helper to state_sync_lib
- shadow_tracking: keep the failover node down for more than the state-sync
  horizon (epoch_sync_horizon_num_epochs, default 2) so on restart it syncs
  state to the current peer-servable epoch, instead of block-syncing and
  catching up an older epoch whose state parts peers no longer serve. Under
  centralized state sync this worked because the external-storage dump was
  epoch-persistent; pure peer sync only serves the current/previous epoch.
- state_sync_lib: move get_external_storage_state_sync_configs() here next to
  the peer-to-peer helpers so state_parts_dump_check only changes its call site.
