+++ pytest/lib/state_sync_lib.py
@@ -1,3 +1,7 @@
+import pathlib
+import tempfile
+
+
 def approximate_epoch_height(block_height, epoch_length):
     if block_height == 0:
         return 0
@@ -52,3 +56,60 @@ def get_state_sync_config_combined():
     }
 
     return config
+
+
+"""
+Generates a pair of configs for centralized (external-storage) state sync.
+    - config_dump: dumps state parts to a local directory.
+    - config_sync: reads state parts from that directory via external storage.
+This is specific to the state-parts dump-check tool, which is external-storage
+only. Prefer the decentralized (peer-to-peer) helpers above for everything else.
+"""
+
+
+def get_external_storage_state_sync_configs():
+    state_parts_dir = str(pathlib.Path(tempfile.gettempdir()) / "state_parts")
+    config_dump = {
+        "state_sync": {
+            "dump": {
+                "location": {
+                    "Filesystem": {
+                        "root_dir": state_parts_dir
+                    }
+                },
+                "iteration_delay": {
+                    "secs": 0,
+                    "nanos": 100000000
+                },
+            }
+        },
+        "store.state_snapshot_config.state_snapshot_type": "Enabled",
+        "tracked_shards_config": 'AllShards'
+    }
+    config_sync = {
+        "consensus.state_sync_external_timeout": {
+            "secs": 0,
+            "nanos": 500000000
+        },
+        "consensus.state_sync_p2p_timeout": {
+            "secs": 0,
+            "nanos": 500000000
+        },
+        "consensus.state_sync_external_backoff": {
+            "secs": 0,
+            "nanos": 500000000
+        },
+        "state_sync": {
+            "sync": {
+                "ExternalStorage": {
+                    "location": {
+                        "Filesystem": {
+                            "root_dir": state_parts_dir
+                        }
+                    }
+                }
+            }
+        },
+        "tracked_shards_config": 'AllShards',
+    }
+    return (config_dump, config_sync)
